♻️  core: Refactor introspect::type module
