+++ zlink-core/src/introspect/type/impls.rs
@@ -1,0 +1,2 @@
+// This file is deprecated and should be removed.
+// The implementations have been moved to the parent directory.
